# Optimize patch search and suggester