--- a/packages/xod-client/src/hinting/actionType.js
+++ b/packages/xod-client/src/hinting/actionType.js
@@ -1,0 +1,1 @@
+export default 'UPDATE_HINTING';