--- a/packages/xod-client/src/hinting/actionTypes.js
+++ b/packages/xod-client/src/hinting/actionTypes.js
@@ -1,6 +1,0 @@
-export const UPDATE_DEDUCED_TYPES = 'UPDATE_DEDUCED_TYPES';
-export const UPDATE_ERRORS = 'UPDATE_ERRORS';
-
-// Just for optimization: call next one once instead of calling two previvous actions
-// For flexibility: keep them
-export const UPDATE_HINTING = 'UPDATE_HINTING';