--- a/packages/xod-client/src/hinting/actions.js
+++ b/packages/xod-client/src/hinting/actions.js
@@ -1,6 +1,6 @@
 import UPDATE_HINTING from './actionType';
 
-export default (deducedTypes, errors) => ({
+export default (deducedTypes, errors, patchSearchData) => ({
   type: UPDATE_HINTING,
-  payload: { deducedTypes, errors },
+  payload: { deducedTypes, errors, patchSearchData },
 });