--- a/packages/xod-client/src/hinting/reducer.js
+++ b/packages/xod-client/src/hinting/reducer.js
@@ -1,4 +1,5 @@
 import * as R from 'ramda';
+import { notNil } from 'xod-func-tools';
 
 import initialState from './state';
 import UPDATE_HINTING from './actionType';
@@ -6,19 +7,35 @@ import { mergeErrors } from './validation';
 
 const errorsLens = R.lensProp('errors');
 
+const getDeducedTypesFromAction = R.path(['payload', 'deducedTypes']);
+const getErrorsFromAction = R.path(['payload', 'errors']);
+const getPatchSearchDataFromAction = R.path(['payload', 'patchSearchData']);
+
 export default (state = initialState, action) => {
   switch (action.type) {
     case UPDATE_HINTING:
       return R.compose(
         R.when(
-          () => action.payload.deducedTypes,
-          R.assoc('deducedTypes', action.payload.deducedTypes)
+          R.pipe(getDeducedTypesFromAction, notNil),
+          R.pipe(
+            getDeducedTypesFromAction,
+            R.assoc('deducedTypes', R.__, state)
+          )
         ),
         R.when(
-          () => action.payload.errors,
-          R.over(errorsLens, mergeErrors(R.__, action.payload.errors))
+          R.pipe(getErrorsFromAction, notNil),
+          R.pipe(getErrorsFromAction, errs =>
+            R.over(errorsLens, mergeErrors(R.__, errs), state)
+          )
+        ),
+        R.when(
+          R.pipe(getPatchSearchDataFromAction, notNil),
+          R.pipe(
+            getPatchSearchDataFromAction,
+            R.assoc('patchSearchData', R.__, state)
+          )
         )
-      )(state);
+      )(action);
     default:
       return state;
   }