--- a/packages/xod-client/src/hinting/reducer.js
+++ b/packages/xod-client/src/hinting/reducer.js
@@ -1,21 +1,23 @@
 import * as R from 'ramda';
 
 import initialState from './state';
-import * as AT from './actionTypes';
+import UPDATE_HINTING from './actionType';
 import { mergeErrors } from './validation';
 
 const errorsLens = R.lensProp('errors');
 
 export default (state = initialState, action) => {
   switch (action.type) {
-    case AT.UPDATE_DEDUCED_TYPES:
-      return R.assoc('deducedTypes', action.payload, state);
-    case AT.UPDATE_ERRORS:
-      return R.over(errorsLens, mergeErrors(R.__, action.payload), state);
-    case AT.UPDATE_HINTING:
+    case UPDATE_HINTING:
       return R.compose(
-        R.assoc('deducedTypes', action.payload.deducedTypes),
-        R.over(errorsLens, mergeErrors(R.__, action.payload.errors))
+        R.when(
+          () => action.payload.deducedTypes,
+          R.assoc('deducedTypes', action.payload.deducedTypes)
+        ),
+        R.when(
+          () => action.payload.errors,
+          R.over(errorsLens, mergeErrors(R.__, action.payload.errors))
+        )
       )(state);
     default:
       return state;