--- a/packages/xod-client/src/hinting/state.js
+++ b/packages/xod-client/src/hinting/state.js
@@ -32,4 +32,15 @@ export default {
     },
     **/
   },
+  patchSearchData: [
+    /**
+     * {
+     *   path: PatchPath,
+     *   lib: String,
+     *   keywords: [String],
+     *   description: String,
+     *   fullDescription: String,
+     * },
+     */
+  ],
 };