There is no issue.

It optimizes the patch search engine by:
- move patch search data into `hinting` branch of the state and update it by middleware, instead of selectors
- update patch search data only on some actions (like `PATCH_DESCRIPTION_UPDATE` or `PROJECT_OPEN`).
- and update search data partially if it possible (updates only patch description, rename patch or add the new one — update only its search data).
- make `searchPatches` function pure and memoize it once
- Suggester component shows only 20 best search results
- Suggester searches when User types at least two symbols

In numbers:
- **Every change** on the patch will not call updating patch search index: ~53ms -> 0ms
- Little changes (description, rename, add): ~53ms -> ~2.5ms
- Change of the entire set of the patches (open/create a project): ~53ms -> ~23.5ms
- Suggester show first results faster on ~100ms